refactor(frontend): split chat controller and runtime helpers (#824)
